+++ src/config/include/config/constants/constants.hpp
@@ -89,6 +89,7 @@ class Constants
     [[nodiscard]] std::filesystem::path getDataPath() const;
     [[nodiscard]] std::filesystem::path getDatabasePath() const;
     [[nodiscard]] std::filesystem::path getImagesPath() const;
+    [[nodiscard]] std::filesystem::path getUpdateStagingPath() const;
 
     [[nodiscard]] static std::string getAppName();
     [[nodiscard]] static std::string getAppDisplayName();